+++ android/app/src/main/kotlin/ca/pigscanfly/opengreeniot/opengreeniot_mobile/MainActivity.kt
@@ -1,0 +1,5 @@
+package ca.pigscanfly.opengreeniot.opengreeniot_mobile
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()